refactor: move the detectors into their own package
The detectors should reside in their own package, at the same level as the issues
