+++ lintrules-gradle/src/main/java/com/chesire/lintrules/gradle/detectors/DependencyDetector.kt
@@ -1,4 +1,4 @@
-package com.chesire.lintrules.gradle
+package com.chesire.lintrules.gradle.detectors
 
 import com.android.tools.lint.detector.api.Detector
 import com.android.tools.lint.detector.api.GradleContext
