+++ lintrules-gradle/src/main/java/com/chesire/lintrules/gradle/issues/DuplicateDependency.kt
@@ -5,7 +5,7 @@ import com.android.tools.lint.detector.api.Implementation
 import com.android.tools.lint.detector.api.Issue
 import com.android.tools.lint.detector.api.Scope
 import com.android.tools.lint.detector.api.Severity
-import com.chesire.lintrules.gradle.DependencyDetector
+import com.chesire.lintrules.gradle.detectors.DependencyDetector
 
 /**
  * Issue for when duplicate dependencies are added to Gradle.
